# storage/framework/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "Docs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        }
    ],
    "paths": {
...
        "/comm/queue": {
            "post": {
                "tags": [
                    "CommController"
                ],
                "summary": "Queue a message to be sent later",
                "operationId": "959834956bcd00e7f5ab93da448d2304",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms",
                                            "whatsapp",
                                            "push"
                                        ],
                                        "example": "email"
                                    },
                                    "to_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com",
                                        "nullable": true
                                    },
                                    "to_phone": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "example": "+919999999999",
                                        "nullable": true
                                    },
                                    "subject": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "Welcome",
                                        "nullable": true
                                    },
                                    "body_text": {
                                        "type": "string",
                                        "example": "Hello from tenant",
                                        "nullable": true
                                    },
                                    "body_html": {
                                        "type": "string",
                                        "example": "<p>Hello from tenant</p>",
                                        "nullable": true
                                    },
                                    "vars_json": {
                                        "type": "object",
                                        "example": {
                                            "name": "Abhay",
                                            "orderId": "12345"
                                        },
                                        "nullable": true,
                                        "additionalProperties": true
                                    },
                                    "template_id": {
                                        "type": "integer",
                                        "example": 12,
                                        "nullable": true
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-09-05T14:30:00+05:30",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 0,
                                        "nullable": true
                                    },
                                    "idempotency_key": {
                                        "type": "string",
                                        "maxLength": 64,
                                        "example": "email-12345-unique",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Queued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "queued": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "outbox_id": {
                                            "type": "integer",
                                            "example": 1024
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The channel field is required."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/t/{tenant}/comm/send": {
            "post": {
                "tags": [
                    "CommController"
                ],
                "summary": "Send a message immediately (tenant -> global provider fallback)",
                "operationId": "edbddff62f947746ccf45f19dee2201f",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms",
                                            "push"
                                        ]
                                    },
                                    "to_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "body_text": {
                                        "type": "string"
                                    },
                                    "body_html": {
                                        "type": "string"
                                    },
                                    "to_phone": {
                                        "type": "string"
                                    },
                                    "to_device_token": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sent"
                    },
                    "500": {
                        "description": "Provider error or misconfiguration"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/t/{tenant}/comm/test": {
            "post": {
                "tags": [
                    "CommController"
                ],
                "summary": "Quick test helper that builds a minimal payload and sends",
                "operationId": "95007dc1280bdf5fdd0ac379d4955424",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel",
                                    "to"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms",
                                            "push"
                                        ]
                                    },
                                    "to": {
                                        "description": "Email, phone, or device token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sent"
                    },
                    "500": {
                        "description": "Provider error or misconfiguration"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}